requirements.txt
python-fasthtml
uvicorn
